+++ hydra-node/src/Hydra/Chain.hs
@@ -4,7 +4,30 @@ module Hydra.Chain where
 
 import Cardano.Prelude
 import Control.Monad.Class.MonadThrow (MonadThrow)
-import Hydra.HeadLogic (OnChainTx)
+import Data.Aeson (FromJSON, ToJSON)
+import Hydra.Ledger (Party, Tx, UTxO)
+import Hydra.Snapshot (Snapshot)
+
+-- NOTE(SN): Might not be symmetric in a real chain client, i.e. posting
+-- transactions could be parameterized using such data types, but they are not
+-- fully recoverable from transactions observed on chain
+-- REVIEW(SN): There is a similarly named type in plutus-ledger, so we might
+-- want to rename this
+data OnChainTx tx
+  = InitTx [Party] -- NOTE(SN): The order of this list is important for leader selection.
+  | CommitTx Party (UTxO tx)
+  | AbortTx (UTxO tx)
+  | CollectComTx (UTxO tx)
+  | CloseTx (Snapshot tx)
+  | ContestTx (Snapshot tx)
+  | FanoutTx (UTxO tx)
+  deriving stock (Generic)
+
+deriving instance Tx tx => Eq (OnChainTx tx)
+deriving instance Tx tx => Show (OnChainTx tx)
+deriving instance Tx tx => Read (OnChainTx tx)
+deriving instance Tx tx => ToJSON (OnChainTx tx)
+deriving instance Tx tx => FromJSON (OnChainTx tx)
 
 data ChainError = ChainError
   deriving (Exception, Show)
