+++ hydra-node/test/Hydra/Chain/ZeroMQSpec.hs
@@ -7,8 +7,8 @@ import Hydra.Prelude
 import Control.Concurrent (newChan, readChan, writeChan)
 import Control.Monad.Class.MonadSTM (newEmptyTMVarIO, putTMVar, takeTMVar)
 import Control.Monad.Class.MonadTimer (timeout)
+import Hydra.Chain (OnChainTx (InitTx))
 import Hydra.Chain.ZeroMQ (catchUpTransactions, mockChainClient, runChainSync, startChain)
-import Hydra.HeadLogic (OnChainTx (InitTx))
 import Hydra.Ledger.Simple (SimpleTx)
 import Hydra.Logging (nullTracer)
 import Test.Hspec.Core.Spec (Spec, describe, it)
